refactor: Consolidate name validation logicc
