+++ src/cargo/util/restricted_names.rs
@@ -1,5 +1,6 @@
 //! Helpers for validating and checking names like package and crate names.
 
+use crate::core::PackageId;
 use crate::util::CargoResult;
 use anyhow::bail;
 use std::path::Path;
@@ -202,4 +203,76 @@ pub fn validate_profile_name(name: &str) -> CargoResult<()> {
     }
 
     Ok(())
+}
+
+pub fn validate_feature_name(pkg_id: PackageId, name: &str) -> CargoResult<()> {
+    if name.is_empty() {
+        bail!("feature name cannot be empty");
+    }
+
+    if name.starts_with("dep:") {
+        bail!("feature named `{name}` is not allowed to start with `dep:`",);
+    }
+    if name.contains('/') {
+        bail!("feature named `{name}` is not allowed to contain slashes",);
+    }
+    let mut chars = name.chars();
+    if let Some(ch) = chars.next() {
+        if !(unicode_xid::UnicodeXID::is_xid_start(ch) || ch == '_' || ch.is_digit(10)) {
+            bail!(
+                "invalid character `{}` in feature `{}` in package {}, \
+                the first character must be a Unicode XID start character or digit \
+                (most letters or `_` or `0` to `9`)",
+                ch,
+                name,
+                pkg_id
+            );
+        }
+    }
+    for ch in chars {
+        if !(unicode_xid::UnicodeXID::is_xid_continue(ch) || ch == '-' || ch == '+' || ch == '.') {
+            bail!(
+                "invalid character `{}` in feature `{}` in package {}, \
+                characters must be Unicode XID characters, '-', `+`, or `.` \
+                (numbers, `+`, `-`, `_`, `.`, or most letters)",
+                ch,
+                name,
+                pkg_id
+            );
+        }
+    }
+    Ok(())
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::sources::CRATES_IO_INDEX;
+    use crate::util::into_url::IntoUrl;
+
+    use crate::core::SourceId;
+
+    #[test]
+    fn valid_feature_names() {
+        let loc = CRATES_IO_INDEX.into_url().unwrap();
+        let source_id = SourceId::for_registry(&loc).unwrap();
+        let pkg_id = PackageId::try_new("foo", "1.0.0", source_id).unwrap();
+
+        assert!(validate_feature_name(pkg_id, "c++17").is_ok());
+        assert!(validate_feature_name(pkg_id, "128bit").is_ok());
+        assert!(validate_feature_name(pkg_id, "_foo").is_ok());
+        assert!(validate_feature_name(pkg_id, "feat-name").is_ok());
+        assert!(validate_feature_name(pkg_id, "feat_name").is_ok());
+        assert!(validate_feature_name(pkg_id, "foo.bar").is_ok());
+
+        assert!(validate_feature_name(pkg_id, "+foo").is_err());
+        assert!(validate_feature_name(pkg_id, "-foo").is_err());
+        assert!(validate_feature_name(pkg_id, ".foo").is_err());
+        assert!(validate_feature_name(pkg_id, "foo:bar").is_err());
+        assert!(validate_feature_name(pkg_id, "foo?").is_err());
+        assert!(validate_feature_name(pkg_id, "?foo").is_err());
+        assert!(validate_feature_name(pkg_id, "ⒶⒷⒸ").is_err());
+        assert!(validate_feature_name(pkg_id, "a¼").is_err());
+        assert!(validate_feature_name(pkg_id, "").is_err());
+    }
 }
